+++ pet_shelter/pet_shelter/src/main/java/com/pet/pet_shelter/DTOs/Document.java
@@ -11,5 +11,5 @@ import lombok.Setter;
 public class Document {
     String path;
     DocType type;
-    int petID;
+    Long petID;
 }
